+++ src/main/java/de/samply/fhir/BundleToContainersConverter.java
@@ -6,6 +6,7 @@ import de.samply.container.Attribute;
 import de.samply.container.Containers;
 import de.samply.converter.ConverterImpl;
 import de.samply.converter.Format;
+import de.samply.exporter.ExporterConst;
 import de.samply.template.AttributeTemplate;
 import de.samply.template.ContainerTemplate;
 import de.samply.template.ConverterTemplate;
@@ -18,6 +19,7 @@ import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.ExpressionNode;
 import org.hl7.fhir.r4.model.Resource;
 import org.hl7.fhir.r4.utils.FHIRPathEngine;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
@@ -25,10 +27,15 @@ import reactor.core.publisher.Flux;
 public class BundleToContainersConverter extends
     ConverterImpl<Bundle, Containers, BundleToContainersConverterSession> {
 
+  private final FhirContext fhirContext;
   private final FHIRPathEngine fhirPathEngine;
+  private final String fhirPackagesDirectory;
 
-  public BundleToContainersConverter() {
-    this.fhirPathEngine = createFhirPathEngine();
+  public BundleToContainersConverter(
+      @Value(ExporterConst.FHIR_PACKAGES_DIRECTORY_SV) String fhirPackagesDirectory) {
+    this.fhirContext = FhirContext.forR4();
+    this.fhirPathEngine = createFhirPathEngine(fhirContext);
+    this.fhirPackagesDirectory = fhirPackagesDirectory;
   }
 
   @Override
@@ -49,13 +56,23 @@ public class BundleToContainersConverter extends
   public Containers convertToContainers(Bundle bundle, ConverterTemplate converterTemplate,
       BundleToContainersConverterSession session) {
     Containers containers = new Containers();
-    BundleContext context = new BundleContext(bundle, session, fhirPathEngine);
+    BundleContext context = createBundleContext(bundle, converterTemplate, session);
     if (converterTemplate != null) {
       converterTemplate.getContainerTemplates()
           .forEach(containerTemplate -> addContainers(bundle, containers, containerTemplate,
               context));
     }
     return containers;
+  }
+
+  private BundleContext createBundleContext(Bundle bundle, ConverterTemplate converterTemplate,
+      BundleToContainersConverterSession session) {
+    try {
+      return new BundleContext(bundle, session, fhirPathEngine, fhirContext, converterTemplate,
+          fhirPackagesDirectory);
+    } catch (BundleContextException e) {
+      throw new RuntimeException(e);
+    }
   }
 
 
@@ -103,11 +120,18 @@ public class BundleToContainersConverter extends
       if (isToBeEvaluated(evalResource, idResource, attributeTemplate)) {
         fhirPathEngine.evaluate(evalResource, expressionNode)
             .forEach(base -> resourceAttributes.add(
-                new ResourceAttribute(idResource, base.toString(), containerTemplate,
-                    attributeTemplate)));
+                new ResourceAttribute(idResource,
+                    fetchAttributeValue(evalResource, attributeTemplate, base, context),
+                    containerTemplate, attributeTemplate)));
       }
     });
     return resourceAttributes;
+  }
+
+  private String fetchAttributeValue(Resource evalResource, AttributeTemplate attributeTemplate,
+      Base base, BundleContext bundleContext) {
+    return (attributeTemplate.isValidation()) ?
+        bundleContext.validate(evalResource, attributeTemplate) : base.toString();
   }
 
   private boolean isToBeEvaluated(Resource evalResource, Resource idResource,
@@ -154,8 +178,7 @@ public class BundleToContainersConverter extends
     return result;
   }
 
-  private FHIRPathEngine createFhirPathEngine() {
-    FhirContext fhirContext = FhirContext.forR4();
+  private FHIRPathEngine createFhirPathEngine(FhirContext fhirContext) {
     return new FHIRPathEngine(
         new HapiWorkerContext(fhirContext, fhirContext.getValidationSupport()));
   }